readme
right click on the omegle chat that you are currently in, click console, click the circle with the line through it to clear console, then insert this code and press enter. this next part is optional, but you can go to an ip lookup site to discover the location of the ip. cities are not always accurate, however if you would like to spook a person, name their county.  the county is more likely to be accurate, however it is not always reliable.